+++ packages/database/src/server-client.ts
@@ -13,7 +13,7 @@ type CookieStore = {
 export function createClient(cookieStore: CookieStore) {
   return createServerClient<Database>(
     process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    process.env.NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY!,
     {
       cookies: {
         getAll() {